# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel and Swagger",
        "description": "Getting started with Laravel and Swagger",
        "termsOfService": "",
        "contact": {
            "email": "name@example.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/walking-classes": {
            "get": {
                "description": "Return Json",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": {
                                "instructor": "A001",
                                "attendent": "22",
                                "vitamin_D": "ddd",
                                "started_at": "20200303",
                                "ended_at": "20200304",
                                "notes": "Check me",
                                "updated_at": "2020-08-26T04:18:51.000000Z",
                                "created_at": "2020-08-26T04:18:51.000000Z",
                                "id": 1
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "description": "Return Json",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "instructor",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "attendent",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "vitamin_D",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "started_at",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ended_at",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "notes",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "examples": {
                            "application/json": {
                                "instructor": "A001",
                                "attendent": "22",
                                "vitamin_D": "ddd",
                                "started_at": "20200303",
                                "ended_at": "20200304",
                                "notes": "Check me",
                                "updated_at": "2020-08-26T04:18:51.000000Z",
                                "created_at": "2020-08-26T04:18:51.000000Z",
                                "id": 1
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/walking-classes/{walkingClassId}": {
            "get": {
                "description": "Return Json",
                "parameters": [
                    {
                        "name": "walkingClassId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "examples": {
                            "application/json": {
                                "instructor": "A001",
                                "attendent": "22",
                                "vitamin_D": "ddd",
                                "started_at": "20200303",
                                "ended_at": "20200304",
                                "notes": "Check me",
                                "updated_at": "2020-08-26T04:18:51.000000Z",
                                "created_at": "2020-08-26T04:18:51.000000Z",
                                "id": 1
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "description": "Return Json",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "walkingClassId",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "instructor",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "attendent",
                        "in": "query",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "vitamin_D",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "started_at",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ended_at",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "notes",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "examples": {
                            "application/json": {
                                "instructor": "A001",
                                "attendent": "22",
                                "vitamin_D": "ddd",
                                "started_at": "20200303",
                                "ended_at": "20200304",
                                "notes": "Check me",
                                "updated_at": "2020-08-26T04:18:51.000000Z",
                                "created_at": "2020-08-26T04:18:51.000000Z",
                                "id": 1
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "description": "Return Json",
                "parameters": [
                    {
                        "name": "walkingClassId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}